+++ distribution/bin/shutdown.sh
@@ -15,14 +15,14 @@
 cd `dirname $0`/../target
 target_dir=`pwd`
 
-pid=`pgrep -f nacos.nacos`
+pid=`pgrep -f nacos-plus.nacos-plus`
 if [ -z "$pid" ] ; then
-        echo "No nacosServer running."
+        echo "No NacosPlus server running."
         exit -1;
 fi
 
-echo "The nacosServer(${pid}) is running..."
+echo "The NacosPlus server(${pid}) is running..."
 
 kill ${pid}
 
-echo "Send shutdown request to nacosServer(${pid}) OK"
+echo "Send shutdown request to NacosPlus server(${pid}) OK"